refactor: move card list into separate component
