+++ src/pages/main-page.tsx
@@ -1,10 +1,10 @@
 import { useState, useEffect } from 'react'
 
-import { CircularProgress, Container } from '@mui/material'
+import { Container } from '@mui/material'
 
 import { BASE_URL } from '../app/constants'
-import { EquipmentCard } from '../components/equipment-card'
-import type { EquipmentID, EquipmentItem } from '../models/equipments'
+import CardList from '../components/equipment-card-list'
+import type { EquipmentItem } from '../models/equipments'
 
 export default function MainPage() {
   const [equipmentList, setEquipmentList] = useState<null | EquipmentItem[]>(null)
@@ -28,44 +28,6 @@ export default function MainPage() {
     fetchEquipmentList()
   }, [])
 
-  function handleClick(e: React.MouseEvent, id: EquipmentID) {} //TODO: implement routing to the card page
-  function handleBtnClick(e: React.MouseEvent, id: EquipmentID) {} //TODO: implement storage data in the favorites
-
-  function renderEquipmentList(): React.ReactNode {
-    if (isLoading) {
-      return <CircularProgress size="60px" />
-    }
-
-    if (isError) {
-      return <h3>Произошла ошибка</h3>
-    }
-
-    if (Array.isArray(equipmentList)) {
-      
-      if (equipmentList.length === 0) {
-        return <h3>Данные отсутствуют</h3>
-      }
-
-      return equipmentList.map(el => {
-        const { id, imgUrl, name, model } = el
-
-        return (
-          <EquipmentCard
-            key={id}
-            id={id}
-            title={name}
-            description={model}
-            imgUrl={imgUrl}
-            handleClick={handleClick}
-            handleBtnClick={handleBtnClick}
-          />
-        )
-      })
-    } else {
-      setIsError(true)
-    }
-  }
-
   return (
     <Container
       sx={{
@@ -76,7 +38,7 @@ export default function MainPage() {
         columnGap: '20px',
       }}
     >
-      {renderEquipmentList()}
+      <CardList list={equipmentList} isLoading={isLoading} isError={isError} />
     </Container>
   )
 }
